Integration tests cleanup
