+++ knotx-it-tests/src/test/resources/conf/routes/te-get.conf
@@ -1,15 +1,10 @@
 routingOperations = ${routingOperations} [
   {
-    operationId = te-local-get
+    operationId = te-get
     handlers = ${config.server.handlers.common.request} [
       {
-        name = fsRepoConnectorHandler
-        config {
-          # Set the root folder of the repository on file system.
-          # If catalogue equals empty string a verticle will look for the files in classpath
-          #
-          catalogue = ""
-        }
+        name = httpRepoConnectorHandler
+        config = ${global.handler.httpRepoConnectorHandler.config}
       },
       {
         name = splitterHandler
