--- knotx-it-tests/src/test/resources/conf/routes/te-remote-get.conf
@@ -1,21 +1,0 @@
-routingOperations = ${routingOperations} [
-  {
-    operationId = te-remote-get
-    handlers = ${config.server.handlers.common.request} [
-      {
-        name = httpRepoConnectorHandler
-        config = ${global.handler.httpRepoConnectorHandler.config}
-      },
-      {
-        name = splitterHandler
-      },
-      {
-        name = fragmentsHandler
-        config = { include required(classpath("conf/routes/handlers/fragmentsHandler.conf")) }
-      },
-      {
-        name = assemblerHandler
-      }
-    ] ${config.server.handlers.common.response}
-  }
-]
