+++ knotx-it-tests/src/test/resources/conf/server.conf
@@ -78,9 +78,7 @@ config.server {
     dropRequests = false
 
     routingOperations = []
-    include required(classpath("conf/routes/te-local-get.conf"))
-    include required(classpath("conf/routes/te-local-post.conf"))
-    include required(classpath("conf/routes/te-remote-get.conf"))
+    include required(classpath("conf/routes/te-get.conf"))
   }
 
   # The options object carries-on configuration called DeploymentOptions for a given verticle.
