--- knotx-it-tests/src/test/resources/content/local/formsBridgeTe.html
@@ -1,77 +1,0 @@
-<!--
-  Copyright (C) 2018 Knot.x Project
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <title>Tempalate with forms, bridge and te</title>
-</head>
-<body>
-  <div>
-    <knotx:snippet knots="form-competition,databridge,te"
-                   databridge-name-labelsrepository="labelsRepository"
-                   databridge-params-labelsrepository='{"path":"/service/mock/labelsRepository.json"}'>
-      <h1>{{labelsrepository._result.welcomeInCompetition}}</h1>
-      {{#if form._result}}
-      <p>{{labelsrepository._result.thankYouForSubscribingToCompetition}}</p>
-      <p>Submitted: {{form._result.form.email}}</p>
-      <p>Submitted: {{form._result.form.name}}</p>
-      {{else}}
-      <p>Please subscribe to our new competition:</p>
-      <form data-knotx-forms-adapter-name="subscribe-competition"
-            data-knotx-forms-on-success="_self"
-            data-knotx-forms-on-error="_self"
-            method="post" class="form-inline">
-        <div class="form-group">
-          <label for="name">Name</label>
-          <input type="text" name="name" id="name"/>
-        </div>
-        <div class="form-group">
-          <label for="email">Email</label>
-          <input type="email" name="email" id="email"/>
-        </div>
-        <button type="submit" class="btn btn-default">Submit</button>
-      </form>
-      {{/if}}
-    </knotx:snippet>
-  </div>
-  <div>
-    <knotx:snippet knots="form-newsletter,databridge,te"
-                   databridge-name-labelsrepository="labelsRepository"
-                   databridge-params-labelsrepository='{"path":"/service/mock/labelsRepository.json"}'>
-      <div id="newsletter-form-wrapper">
-        <h1>{{labelsrepository._result.subscribeToNewsletter}}</h1>
-        {{#if form._result}}
-        <p>{{labelsrepository._result.thankYouForSubscribingToNewsletter}}</p>
-        <p>Submitted: {{form._result.form.email}}</p>
-        {{else}}
-        <p>Please subscribe to our newsletter:</p>
-        <form data-knotx-forms-adapter-name="subscribe-newsletter"
-              data-knotx-forms-on-success="_self"
-              data-knotx-forms-on-error="_self"
-              method="post" class="form-inline">
-          <div class="form-group">
-            <label for="email2">Email</label>
-            <input type="email" name="email" id="email2"/>
-          </div>
-          <button type="submit" class="btn btn-default">Submit</button>
-        </form>
-        {{/if}}
-      </div>
-    </knotx:snippet>
-  </div>
-</body>
-</html>
