--- knotx-it-tests/src/test/resources/content/local/notExistingDataDefinitionWithFallback.html
@@ -1,49 +1,0 @@
-<!--
-  Copyright (C) 2018 Knot.x Project
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <title>Missing service</title>
-</head>
-<body>
-  <div>
-    <knotx:fallback fallback-id="CUSTOM">
-      <p class="error">local fallback</p>
-    </knotx:fallback>
-    <knotx:snippet knots="databridge,te"
-                   databridge-name="first-service"
-                   databridge-name-undefined="non-existing-service"
-                   fallback="CUSTOM"
-                   type="text/knotx-snippet">
-      <h2>Things below will be replaced by fallback markup</h2>
-      <div>
-        <h2>Snippet1 - {{_result.message}}</h2>
-        <div>Snippet1 - {{_result.body.a}}</div>
-      </div>
-    </knotx:snippet>
-  </div>
-  <div>
-    <knotx:snippet knots="databridge,te"
-                   databridge-name="authors"
-                   type="text/knotx-snippet">
-      <h2>This should work when global fallback defined</h2>
-      <h2>{{_result.name}}</h2>
-      <p>Bio: {{_result.info.bio}}</p>
-    </knotx:snippet>
-  </div>
-</body>
-</html>
