+++ knotx-it-tests/src/test/resources/results/fullPage.html
@@ -1,5 +1,5 @@
 <!--
-  Copyright (C) 2018 Knot.x Project
+  Copyright (C) 2019 Knot.x Project
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -21,7 +21,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="shortcut icon" href="http://cognifide.com/favicon.ico?v=2"/>
   <title>Knot.x</title>
-  <link href="https://bootswatch.com/simplex/bootstrap.min.css" rel="stylesheet"/>
+  <link href="https://bootswatch.com/superhero/bootstrap.min.css" rel="stylesheet"/>
 </head>
 <body>
   <div class="container-fluid">
