--- knotx-it-tests/src/test/resources/results/remote-fullPage.html
@@ -1,71 +1,0 @@
-<!--
-  Copyright (C) 2018 Knot.x Project
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="shortcut icon" href="http://cognifide.com/favicon.ico?v=2"/>
-  <title>Knot.x</title>
-  <link href="https://bootswatch.com/superhero/bootstrap.min.css" rel="stylesheet"/>
-</head>
-<body>
-  <div class="container-fluid">
-    <div class="row">
-      <div class="col-md-12">
-        <div class="jumbotron">
-          <h2> Hello Knot.x - reactive micro-service assembler world! </h2>
-          <p> This template is served from the <strong>local</strong> repository. </p>
-          <p> <a class="btn btn-primary btn-large" href="https://github.com/Cognifide/knotx">Learn more</a> </p>
-        </div>
-      </div>
-    </div>
-    <div class="row">
-      <div class="col-md-4">
-        <h2>Book</h2>
-        <p>Knot.x Integration Tests in Practice</p>
-        <p>Authors: Knot.x Team</p>
-        <div>Success! Status code : 200</div>
-      </div>
-      <div class="col-md-4">
-        <h2>Author</h2>
-        <p>Knot X</p>
-        <p>Bio: Aliquam cursus fermentum mi, vel tempus mi pellentesque vel. Nunc in feugiat lorem. Etiam placerat ante eget sem euismod, sit amet dictum sapien vulputate. Morbi pellentesque arcu a mauris ornare eleifend.</p>
-        <div>Success! Status code : 200</div>
-      </div>
-    </div>
-    <div class="row">
-      <div class="col-md-4">
-        <h2>Portfolio: </h2>
-        <ul>
-          <li>tutorials</li>
-          <li>blogs</li>
-          <li>books</li>
-        </ul>
-      </div>
-      <div class="col-md-4">
-        <h2>Tags: </h2>
-        <ul>
-          <li>reactive</li>
-          <li>integration</li>
-          <li>framework</li>
-        </ul>
-      </div>
-    </div>
-  </div>
-</body>
-</html>
