--- knotx-it-tests/src/test/resources/results/submitNewsletterForm.html
@@ -1,47 +1,0 @@
-<!--
-  Copyright (C) 2018 Knot.x Project
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <title>After newsletter form submitted</title>
-</head>
-<body>
-  <div>
-    <h1>welcome in competition</h1>
-    <p>Please subscribe to our new competition:</p>
-    <form method="post" class="form-inline">
-      <input type="hidden" name="_frmId" value="competition" />
-      <div class="form-group">
-        <label for="name">Name</label>
-        <input type="text" name="name" id="name" />
-      </div>
-      <div class="form-group">
-        <label for="email">Email</label>
-        <input type="email" name="email" id="email" />
-      </div>
-      <button type="submit" class="btn btn-default">Submit</button>
-    </form>
-  </div>
-  <div>
-    <div id="newsletter-form-wrapper">
-      <h1>subscribe to newsletter</h1>
-      <p>thank you for subscribing to newsletter</p>
-      <p>Submitted: email-2@example.com</p>
-    </div>
-  </div>
-</body>
-</html>
